Fixed hovering for tokens
The problem was that the boundingRect of the tokenCluster is larger than the token and would go in front of the tokens when it detects a hoverEvent. Fixed by disabling hoverEvents for the tokenCluster when there are less than 4 tokens.

Also reversed the order in which the tokens are drawn, so they are now the wrong way around :)

Sometimes the program crashes when messing with the values of the tokens, doesn't give an error, so not sure why.